pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yagpt"
version = "0.2.0"
description = "Yet Another GPT - A clean, modular GPT implementation for learning and experimentation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["gpt", "transformer", "language-model", "pytorch", "deep-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "tiktoken>=0.5.0",
    "pyarrow>=14.0.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]
logging = [
    "wandb>=0.15.0",
]
eval = [
    "lm-eval>=0.4.0",
]
all = [
    "yagpt[dev,logging,eval]",
]

[project.scripts]
yagpt = "scripts.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["yagpt*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = ["-ra", "--strict-markers"]
